move NFTListRow to its own file
